requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
dnspython==1.16.0
exceptiongroup==1.2.0; python_version < '3.11'
fastapi==0.110.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
pydantic==2.6.4; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
pymongo[srv]==3.6.0
python-dotenv==1.0.1; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
typing-extensions==4.10.0; python_version >= '3.8'
uvicorn==0.29.0; python_version >= '3.8'
